+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added new module `typing.py` for custom types
 - Added new module `enums.py` and support of enums in arguments of methods
 - Added `category_of_website` and `port` to `Internet` data provider
+- Added `price_in_btc` to `Business` data provider
 - Added `dna`, `rna` and `atomic_number` to `Science` data provider
 - Added `schoice` method for `Random`
 - Added alias `last_name` for `surname` in `Personal` data provider